pyproject.toml
[tool.poetry]
name = "csv-cleaner"
version = "0.1.0"
description = "CSV Cleaner Demo"
authors = ["Duy Nguyen"]
license = "OpenGL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
django = "^5.0.3"
environs = "^11.0.0"
djangorestframework = "^3.15.0"
python-json-logger = "^2.0.7"
djangorestframework-dataclasses = "^1.3.1"
djangorestframework-camel-case = "^1.4.2"
pytest = "^8.1.1"
pylint = "^3.1.0"
pylint-django = "^2.5.5"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
